Read Horiba Labspec files (ASCII/txt)

Read ASCII (.txt) files exported by Horiba's Labspec software (LabRAM spectrometers).

read_txt_Horiba(
  file,
  cols = c(spc = "I / a.u.", .wavelength = expression(Delta * tilde(nu)/cm^-1)),
  header = TRUE,
  sep = "\t",
  row.names = NULL,
  check.names = FALSE,
  ...
)

read_txt_Horiba_xy(file, ...)

read_txt_Horiba_t(
  file,
  header = TRUE,
  sep = "\t",
  row.names = NULL,
  check.names = FALSE,
  ...
)

Arguments

file

Path of connection to .txt file.

cols, header, sep, row.names, check.names, ...

further parameters are handed over to hyperSpec::read_txt_wide()

Value

hyperSpec object.

Details

read_txt_Horiba()

read_txt_Horiba_xy() reads maps, i.e. .txt files where the first two columns give x and y coordinates.

read_txt_Horiba_t() reads time series, i.e. .txt files with the time in the first column.

Author

C. Beleites