Read WITec files (ASCII/txt)

Import Raman spectra/maps from WITec Instrument via ASCII files

read_dat_WITec(
  filex = stop("filename or connection needed"),
  filey = sub("-x", "-y", filex),
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = hy_get_option("debuglevel") < 1L
)

read_txt_WITec_Graph(
  headerfile = stop("filename or connection needed"),
  filex = gsub("Header", "X-Axis", headerfile),
  filey = gsub("Header", "Y-Axis", headerfile),
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = TRUE
)

read_txt_WITec(
  file = stop("filename or connection needed"),
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  hdr.label = FALSE,
  hdr.units = FALSE,
  encoding = "unknown",
  ...,
  quiet = TRUE
)

Arguments

filex

filename wavelength axis file

filey

filename intensity file

points.per.line

Number of spectra in x direction of the map.

lines.per.image

Number of spectra in y direction.

type

Type of spectra:

single for single spectra (including time series);

map for imaging data.

encoding

character encoding, see base::readLines().

..., quiet

handed to base::scan().

headerfile

Path or connection to ASCII file with header information.

file

Path or connection to ASCII file.

hdr.label

WITec Project exports the spectra names (contain information of map position or number of spectra) within the file.

hdr.units

WITec Project exports the spectra units within the file.

Value

hyperSpec object.

Details

read_txt_WITec() reads WITec ASCII files where the first column gives the wavelength axes and the other columns the spectra.

read_dat_WITec() reads WITec's ASCII exported data which comes in separate files with x and y data.

Author

Claudia Beleites and Marcel Dahms